move all code to cambricon/collective_communication
